requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
pydantic>=2.9.0
supabase==2.15.0
python-dotenv==1.0.1
httpx>=0.24.1,<0.29.0
svix>=1.15.0